[difftest] fix QEMU side difftest

+++ include/dut.h
@@ -5,8 +5,8 @@
 #include "qemu.h"
 
 typedef struct {
-    int mycpu_pcs[3];
-}  diff_pcs;
+    uint64_t mycpu_pcs[3];
+} diff_pcs;
 
 // TODO sync cycle and sync interrupt
 void dut_reset(int cycle, char *elf_path);  // reset processor and initialize memory
